requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
goodfire>=0.1.0
colorama>=0.4.6 